python_mcp_server/requirements.txt
mcp
fastapi
uvicorn
PyYAML
requests
